switched to doubles for larger vectors and values

+++ vec_multiply/VectorMultiply.cu
@@ -19,7 +19,7 @@ void cudaCheckError(cudaError_t err) {
 }
 
 __global__
-void vecMultKernel(float * a, float * b, float * c, int size)
+void vecMultKernel(double * a, double * b, double * c, int size)
 {
   int i = blockDim.x*blockIdx.x+threadIdx.x;
 
@@ -27,14 +27,14 @@ void vecMultKernel(float * a, float * b, float * c, int size)
 
 }
 
-float vecMultGPU(float * h_a, float * h_b, float * h_c, int len, CudaBlockInfo * blockInfo){
+double vecMultGPU(double * h_a, double * h_b, double * h_c, int len, CudaBlockInfo * blockInfo){
 
 	cudaEvent_t start, stop;
 	float elapsedTime;
 
-  int size = len * sizeof(float);
-  float * d_a, * d_b, * d_c;
-  int result = 0;
+  int size = len * sizeof(double);
+  double * d_a, * d_b, * d_c;
+  double result = 0;
 
   // alocate memory and move vectors to GPU
   cudaCheckError(cudaMalloc((void**)&d_a, size));
@@ -74,10 +74,10 @@ float vecMultGPU(float * h_a, float * h_b, float * h_c, int len, CudaBlockInfo *
   return result;
 }
 
-float vecMultCPU(float * a, float * b, float * c, int len){
+double vecMultCPU(double * a, double * b, double * c, int len){
 	cudaEvent_t start, stop;
 	float elapsedTime;
-  int result = 0;
+  double result = 0;
 
 	cudaEventCreate(&start);    
 	cudaEventCreate(&stop);
@@ -101,14 +101,14 @@ float vecMultCPU(float * a, float * b, float * c, int len){
   return result;
 }
 
-void fillVecs(float * a, float * b, int size){
+void fillVecs(double * a, double * b, int size){
   for(int i=0; i<size; ++i){
     a[i] = i;
     b[i] = i;
   }
 }
 
-void printVecs(float * a, float * b, float * c, int len){
+void printVecs(double * a, double * b, double * c, int len){
   for(int i=0; i<len; ++i){
     printf("a:%f b:%f, c:%f\n",a[i], b[i], c[i]);
   }
@@ -129,7 +129,7 @@ int cudaDeviceProperties(){
     return 0;
   }
 
-  float bytesInGiB = 1 << 30;
+  double bytesInGiB = 1 << 30;
 
   // print stats for each cuda device found
   for (int i = 0; i < nDevices; ++i){
@@ -167,10 +167,10 @@ int validateBlockInfoForDevice(CudaBlockInfo  * blockInfo, int vecLen, int devic
       printf("  Max blocks per grid: %d\n", prop.maxGridSize[0]);
       printf("  Requested blocks per grid: %d\n", blockInfo->blocksPerGrid);
     }
-    else if(prop.totalGlobalMem < 3*(vecLen*sizeof(float))){
+    else if(prop.totalGlobalMem < 3*(vecLen*sizeof(double))){
       printf("\nDevice %s is unable to process request!\n", prop.name);
       printf("  Total global memory is: %lu\n", prop.totalGlobalMem);
-      printf("  Bytes needed for vectors: %lu\n", 3*(vecLen*sizeof(float)));
+      printf("  Bytes needed for vectors: %lu\n", 3*(vecLen*sizeof(double)));
     }
     else{
       return 1;
@@ -235,10 +235,10 @@ int main(int argc, char * argv[])
   }
 
   // vector variables
-  float * a = (float*)malloc(*vecLength * sizeof(float));
-  float * b = (float*)malloc(*vecLength * sizeof(float));
-  float * c = (float*)malloc(*vecLength * sizeof(float));
-  float result = 0;
+  double * a = (double*)malloc(*vecLength * sizeof(double));
+  double * b = (double*)malloc(*vecLength * sizeof(double));
+  double * c = (double*)malloc(*vecLength * sizeof(double));
+  double result = 0;
 
   // fill vectors a and b with values
   fillVecs(a, b, *vecLength);
@@ -246,14 +246,14 @@ int main(int argc, char * argv[])
   printf("\nVector multiplication using CPU with %d elements:\n", *vecLength);
 
   result = vecMultCPU(a, b, c, *vecLength);
-  printf("Result of vector multiplication on the CPU: %.2f\n", result);
+  printf("Result of vector multiplication on the CPU: %.2lf\n", result);
   //printVecs(a, b, c, *vecLength);
 
   printf("\nVector multiplication using GPU with %d elements, %d threads per block and %d blocks per grid:\n", 
          *vecLength, blockInfo->threadsPerBlock, blockInfo->blocksPerGrid);
 
   result = vecMultGPU(a, b, c, *vecLength, blockInfo);
-  printf("Result of vector multiplication on the GPU: %.2f\n", result);
+  printf("Result of vector multiplication on the GPU: %.2lf\n", result);
   //printVecs(a, b, c, *vecLength);
 
   // free memory
